Add Ecto/Postgres dependency.
Use `mix deps.get` to resolve and download dependencies.

Running tests with `mix test` or loading the app with `iex -S mix` will
compile the dependencies.

The `elixir_apps` definition is a list of what other OTP applications
need to be started. In this case postgres and ecto will start a series
of processes to manage postgres connections and handle requests.